rlib_fft
Module precision

Constants

CORRECT_F32_BOUNDS

CORRECT_F64_BOUNDS

For each pair of (A, B) from VALS_TO_CHECK this array stores (more or less maximum possible) value L such that multiplying array a with values [0..=A] of len L with array b with values [0..=B] of len L will give correct result.

VALS_TO_CHECK